# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  internal_server.AccountLoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: password123{#Pbb
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_server.AccountLoginResponse:
    properties:
      access_token:
        description: SessionID             uuid.UUID    `json:"session_id" example:"123e4567-e89b-12d3-a456-426614174000"`
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        type: string
      access_token_expires_at:
        example: "2025-02-05T13:15:08Z"
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        type: string
      refresh_token_expires_at:
        example: "2025-02-06T13:15:08Z"
        type: string
      user:
        $ref: '#/definitions/internal_server.AccountResponse'
    type: object
...
  internal_server.AccountResponse:
    properties:
      created_at:
        example: "2025-01-01T12:00:00Z"
        type: string
      department:
        example: Minerals Department
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John_doe11
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      last_active:
        example: "2025-01-02T12:00:00Z"
        type: string
      last_name:
        example: John_doe11
        type: string
      role:
        example: ADMIN
        type: string
      status:
        example: ACTIVE
        type: string
      updated_at:
        example: "2025-01-02T12:00:00Z"
        type: string
      user_name:
        example: John_doe142
        type: string
    type: object
...
  internal_server.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  internal_server.Message:
    properties:
      code:
        example: 500
        type: integer
      error:
        example: Failed to Process Task
        type: string
      message:
        example: Task status updated - COMPLETED
        type: string
      result:
        example: Task processed successfully
        type: string
      task_id:
        example: 1233-flf4djf-alsdik
        type: string
    type: object
...
  internal_server.SendOTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  internal_server.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  internal_server.renewAccessTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_server.renewAccessTokenResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the github.com/0xdbb/eggsplore service
  title: github.com/0xdbb/eggsplore API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login account with email and password
      parameters:
      - description: Account Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server.AccountLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      summary: Login Account
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout account by deleting session and clearing cookies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      summary: Logout Account
      tags:
      - auth
  /auth/renew:
    post:
      consumes:
      - application/json
      description: Generates a new access token using a valid refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_server.renewAccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.renewAccessTokenResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "401":
          description: Unauthorized or Invalid token
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      summary: Renew Access Token
      tags:
      - auth
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: Resend OTP to a verified account's email if previous OTP is expired
      parameters:
      - description: Send OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server.SendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP resent successfully
          schema:
            $ref: '#/definitions/internal_server.Message'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "401":
          description: Email not verified or OTP not expired
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "429":
          description: OTP still valid, wait before requesting new one
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      summary: Resend OTP
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies the OTP sent to the account's email and generates access
        and refresh tokens
      parameters:
      - description: OTP Verification Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful verification and login
          schema:
            $ref: '#/definitions/internal_server.AccountLoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "401":
          description: Invalid or expired OTP
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Invalid email or OTP
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      summary: Verify 2FA OTP to complete login
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"